+++ crates/ytsaurus-job/src/lib.rs
@@ -66,11 +66,14 @@
 pub mod error;
 /// Reading the input stream.
 pub mod reader;
+/// Custom job statistics.
+pub mod statistics;
 /// Writing output tables.
 pub mod writer;
 
 pub use crate::error::{JobError, Result};
 pub use crate::reader::{Event, Group, GroupKey, Groups, JobReader, Row};
+pub use crate::statistics::JobStatistics;
 pub use crate::writer::{JobWriter, TableId, table_descriptor};
 
 pub use ytsaurus_yson as yson;